Moved XSS checks into its own class
